remove test exclusions and add reusable component
