+++ frontend/.scannerwork/report-task.txt
@@ -2,5 +2,5 @@ projectKey=POC-frontend
 serverUrl=http://localhost:9000
 serverVersion=9.7.0.61563
 dashboardUrl=http://localhost:9000/dashboard?id=POC-frontend
-ceTaskId=AYRXwTA91-pyWhb0kykB
-ceTaskUrl=http://localhost:9000/api/ce/task?id=AYRXwTA91-pyWhb0kykB
+ceTaskId=AYRaRvBY1-pyWhb0kykE
+ceTaskUrl=http://localhost:9000/api/ce/task?id=AYRaRvBY1-pyWhb0kykE
