+++ frontend/sonar-project.properties
@@ -2,7 +2,8 @@ sonar.projectKey=POC-frontend
 sonar.javascript.lcov.reportPaths=./coverage/lcov.info
 sonar.login=sqp_05efbce5a48ef4e8460ed2417261085ea969d9dc
 sonar.sources=src
-sonar.coverage.exclusions=**/test-coverage/**, **/coverage/**, /src/utils/*, /src/features/**/*, /src/hooks/*, /src/App.tsx, /src/index.ts, /src/react-app-env.d.ts, /src/reportWebVitals.ts, src/@types/*, /src/components/index.ts, /src/mocks/*, /src/pages/index.ts, /src/styles/*, /src/test/**, /src/index.tsx
+sonar.coverage.exclusions=**/coverage/**, /src/test/**
 sonar.tests.inclusions=/src/test/**
+sonar.cpd.exclusions=/src/test/**
 sonar.language=ts
 sonar.sourceEncoding=UTF-8
