+++ frontend/src/components/Tables/TableMovies.tsx
@@ -5,7 +5,6 @@ import {
   MultiSelect,
   NumberInput,
   Paper,
-  SegmentedControl,
   Space,
   Text,
   Textarea,
@@ -30,7 +29,6 @@ import {
   IMovie,
   IMovieReview,
   IPatchMovie,
-  IPostActor,
   IPostMovie,
 } from "../../utils/types";
 import { tableCustomStyles, useStyles } from "./TableStyles";
@@ -41,14 +39,13 @@ import {
   fetchMovieReviewsById,
   updateMovieById,
 } from "../../features/movie/movieSlice";
-import { addActor } from "../../features/actor/actorSlice";
 import {
   budgetFormatter,
   isNotEmpty,
-  isValidName,
   isValidUrl,
   movieRating,
 } from "../../utils/helpers";
+import AddActorModal from "../AddActorModal.tsx/AddActorModal";
 
 const TableMovies = () => {
   const { movies, isLoading } = useTypedSelector((state) => state.movie);
@@ -76,14 +73,6 @@ const TableMovies = () => {
     yearReleased: 2022,
     image: "",
     actors: [],
-  });
-  const [newActor, setNewActor] = useState<IPostActor>({
-    firstName: "",
-    lastName: "",
-    gender: "male",
-    age: 0,
-    image: "",
-    link: "",
   });
   const [selectedMovieData, setSelectedMovieData] = useState<IPatchMovie>(
     {} as IPatchMovie
@@ -140,41 +129,6 @@ const TableMovies = () => {
       });
     }
   }, [dispatch, newMovie]);
-
-  // Add Actor Action (POST request)
-  const handleAddActor = useCallback(async () => {
-    try {
-      // Validate input fields
-      isValidName(newActor.firstName, "first");
-      isValidName(newActor.lastName, "last");
-      if (newActor.gender !== "male" && newActor.gender !== "female")
-        throw new Error("Gender should only be either male or female");
-      if (newActor.age < 1 || !newActor.age)
-        throw new Error("Actor age cannot be less than a year.");
-      isValidUrl(newActor.image, "actor image");
-      isValidUrl(newActor.link as string, "actor link");
-
-      const res: IDispatchResponse = await dispatch(addActor(newActor));
-      if (!res.error) {
-        setAddActorModal(false);
-        setNewActor({
-          firstName: "",
-          lastName: "",
-          gender: "male",
-          age: 0,
-          image: "",
-          link: "",
-        });
-      }
-    } catch (error: any) {
-      showNotification({
-        title: "Adding actor failed. See message below for more info.",
-        message: error.message,
-        autoClose: 3000,
-        color: "yellow",
-      });
-    }
-  }, [dispatch, newActor]);
 
   // Open update modal and set current row item data to selectedMovieData state
   const handleMovieUpdateActionClick = useCallback(
@@ -238,12 +192,6 @@ const TableMovies = () => {
     {
       name: "Title",
       selector: (row) => upperFirst(row.title),
-      sortable: true,
-      reorder: true,
-    },
-    {
-      name: "Description",
-      selector: (row) => upperFirst(row.description),
       sortable: true,
       reorder: true,
     },
@@ -442,83 +390,10 @@ const TableMovies = () => {
       </Modal>
 
       {/* Add Actor Modal */}
-      <Modal
-        opened={addActorModal}
-        onClose={() => setAddActorModal(false)}
-        title="Add Actor"
-        centered
-      >
-        <TextInput
-          placeholder="First Name"
-          label="First Name"
-          defaultValue={newActor?.firstName}
-          onChange={(e) =>
-            setNewActor({ ...newActor, firstName: e.currentTarget.value })
-          }
-          withAsterisk
-        />
-        <TextInput
-          placeholder="Last Name"
-          label="Last Name"
-          defaultValue={newActor?.lastName}
-          onChange={(e) =>
-            setNewActor({ ...newActor, lastName: e.currentTarget.value })
-          }
-          withAsterisk
-        />
-        <NumberInput
-          placeholder="Actor age"
-          label="Actor age"
-          defaultValue={newActor?.age}
-          onChange={(value) =>
-            setNewActor({ ...newActor, age: value as number })
-          }
-          hideControls
-          withAsterisk
-        />
-        <TextInput
-          placeholder="Actor link (e.g. IMDB or Wikipedia)"
-          label="Actor Link"
-          defaultValue={newActor?.link}
-          onChange={(e) =>
-            setNewActor({ ...newActor, link: e.currentTarget.value })
-          }
-        />
-        <TextInput
-          placeholder="Actor Image URL"
-          label="Actor Image"
-          defaultValue={newActor?.image}
-          onChange={(e) =>
-            setNewActor({ ...newActor, image: e.currentTarget.value })
-          }
-          withAsterisk
-        />
-
-        <SegmentedControl
-          defaultValue={newActor?.gender}
-          mt="15px"
-          onChange={(value) =>
-            setNewActor({
-              ...newActor,
-              gender: value as "male" | "female",
-            })
-          }
-          data={[
-            { label: "Gender", value: "gender", disabled: true },
-            { label: "Male", value: "male" },
-            { label: "Female", value: "female" },
-          ]}
-        />
-
-        <Button
-          className={classes.btn}
-          size="xs"
-          mt="lg"
-          onClick={handleAddActor}
-        >
-          Add Actor
-        </Button>
-      </Modal>
+      <AddActorModal
+        addActorModal={addActorModal}
+        setAddActorModal={setAddActorModal}
+      />
 
       {/* Edit Movie Modal */}
       <Modal
